+++ crates/env/src/call/common.rs
@@ -18,13 +18,16 @@ use core::marker::PhantomData;
 
 use ink_primitives::{
     abi::{
-        AbiDecodeWith,
         Ink,
         Sol,
     },
+    sol::{
+        SolResultDecode,
+        SolResultDecodeError,
+    },
     MessageResult,
-    SolDecode,
 };
+use pallet_revive_uapi::ReturnErrorCode;
 use scale::{
     Decode,
     DecodeAll,
@@ -128,7 +131,10 @@ impl<T> Unwrap for Set<T> {
 pub trait DecodeMessageResult<Abi>: Sized {
     /// Decodes the output of a message call, requiring the output
     /// to be wrapped with `MessageResult` (if not included in the output).
-    fn decode_output(buffer: &[u8]) -> crate::Result<MessageResult<Self>>;
+    fn decode_output(
+        buffer: &[u8],
+        did_revert: bool,
+    ) -> crate::Result<MessageResult<Self>>;
 }
 
 impl<R> DecodeMessageResult<Ink> for R
@@ -136,7 +142,7 @@ where
     R: Decode,
     MessageResult<R>: Decode,
 {
-    fn decode_output(mut buffer: &[u8]) -> crate::Result<MessageResult<Self>> {
+    fn decode_output(mut buffer: &[u8], _: bool) -> crate::Result<MessageResult<Self>> {
         let decoded = MessageResult::<R>::decode_all(&mut buffer)?;
         Ok(decoded)
     }
@@ -144,12 +150,26 @@ where
 
 impl<R> DecodeMessageResult<Sol> for R
 where
-    R: SolDecode,
+    R: SolResultDecode,
 {
-    fn decode_output(buffer: &[u8]) -> crate::Result<MessageResult<Self>> {
+    fn decode_output(
+        buffer: &[u8],
+        did_revert: bool,
+    ) -> crate::Result<MessageResult<Self>> {
         // Solidity ABI Encoded contracts return the data without
         // `MessageResult`.
-        let decoded = R::decode_with(buffer)?;
+        let decoded = R::decode(buffer, did_revert)?;
         Ok(Ok(decoded))
+    }
+}
+
+impl From<SolResultDecodeError> for crate::Error {
+    fn from(value: SolResultDecodeError) -> Self {
+        match value {
+            SolResultDecodeError::NonResultFromRevert => {
+                Self::ReturnError(ReturnErrorCode::CalleeReverted)
+            }
+            SolResultDecodeError::Decode => Self::DecodeSol(ink_primitives::sol::Error),
+        }
     }
 }